requirements.txt
langchain>=0.1.0
faiss-cpu>=1.7.4
transformers>=4.35.0
torch>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pypdf>=3.17.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pytest>=7.3.1
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.23.0
streamlit>=1.28.0
sentence-transformers>=2.2.2
datasets>=2.14.0